+++ src/Sdk/DTWebApi/WebApi/RunnerRefreshMessage.cs
@@ -11,36 +11,6 @@ namespace GitHub.DistributedTask.WebApi
     {
         public static readonly String MessageType = "RunnerRefresh";
 
-        [DataContract]
-        public sealed class BrokerPackageMetadata
-        {
-            [JsonConstructor]
-            internal BrokerPackageMetadata()
-            {
-            }
-
-            [DataMember(Name = "download_url")]
-            public string DownloadUrl
-            {
-                get;
-                set;
-            }
-
-            [DataMember(Name = "sha256_checksum")]
-            public string HashValue
-            {
-                get;
-                set;
-            }
-
-            [DataMember(Name = "os")]
-            public string Platform
-            {
-                get;
-                set;
-            }
-        }
-
         [JsonConstructor]
         internal RunnerRefreshMessage()
         {
